requirements.txt
asgiref==3.6.0
Django==4.2
django-allauth==0.45.0
django-ckeditor==6.7.0
django-crispy-forms==1.9.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-solo==2.1.0
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
Pillow==10.0.0
psycopg2-binary==2.9.5
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
